requirements.txt
a-pytube-fork-for-spotdl-users==9.6.4
aenum==2.2.3
altgraph==0.17
amqp==2.6.1
app-version==1.0.1
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
Django==2.2.13
django-audiofield==0.10.0
django-bootstrap-modal-forms==1.4.1
django-celery==3.3.1
django-crispy-forms==1.11.0
django-smart-selects==1.5.9
django-widget-tweaks==1.4.2
fbchat==1.9.6
fdfgen==0.16.1
filelock==3.0.12
future==0.17.1
gdown==3.11.0
gkeepapi==0.10.9
gpsoauth==0.4.1
idna==2.8
importlib-metadata==1.6.0
isort==4.3.21
joblib==0.15.1
keepcli==1.0.1
keyboard==0.13.5
keyring==21.2.1
kiwisolver==1.2.0
kombu==4.6.11
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
mutagen==1.45.1
nltk==3.5
nmap==0.0.1
nose==1.3.7
notify==0.3.1
notify-run==0.0.13
numpy==1.18.4
packaging==20.4
paho-mqtt==1.5.0
pdfrw==0.4
pefile==2019.4.18
Pillow==7.1.2
punctuator==0.9.5
PyAudio==0.2.11
pycparser==2.20
pycryptodomex==3.9.7
pyglet==1.4.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pylint==2.5.3
pyparsing==2.4.7
PyPDF2==1.26.0
pypdftk==0.4
PyQRCode==1.2.1
PySocks==1.7.1
python-dateutil==2.8.1
python-nmap==0.6.1
pytz==2018.5
pywin32-ctypes==0.2.0
PyYAML==3.13
rapidfuzz==0.14.1
regex==2020.5.14
reportlab==3.5.53
requests==2.22.0
rope==0.18.0
scipy==1.4.1
selenium==3.14.0
six==1.15.0
sounddevice==0.3.15
soupsieve==2.0.1
SpeechRecognition==3.8.1
spotdl==3.1.4
spotipy==2.16.1
sqlparse==0.3.0
termcolor==1.1.0
Theano==1.0.4
toml==0.10.1
tqdm==4.46.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.9
validate-email==1.3
vine==1.3.0
virtualenv==16.7.8
Wave==0.0.2
wrapt==1.12.1
zipp==3.1.0
